adding transactions for save garbage

+++ src/main/java/com/co/latin/ecociudadano/repositories/ContainerRepository.java
@@ -3,5 +3,5 @@ package com.co.latin.ecociudadano.repositories;
 import com.co.latin.ecociudadano.model.Container;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ContainerRepository extends CrudRepository<Container, Integer> {
+public interface ContainerRepository extends CrudRepository<Container, Long> {
 }
